chore: move PartialAssignment object to update package as util, update relative tests
