+++ src/main/scala/satify/update/solver/dpll/DpllDecision.scala
@@ -2,10 +2,10 @@ package satify.update.solver.dpll
 
 import satify.model.cnf.Bool.False
 import satify.model.cnf.CNF.Symbol
-import satify.model.dpll.PartialAssignment.{filterUnconstrVars, updatePartialAssignment}
 import satify.model.dpll.{Constraint, Decision, OptionalVariable}
 import satify.update.solver.dpll.Optimizations.{pureLiteralIdentification, unitLiteralIdentification}
 import satify.update.solver.dpll.cnf.CNFSimplification.simplifyCnf
+import satify.update.solver.dpll.utils.PartialAssignmentUtils.{filterUnconstrVars, updatePartialAssignment}
 
 import scala.util.Random
