+++ src/main/scala/satify/update/solver/dpll/impl/DpllEnumerator.scala
@@ -4,12 +4,13 @@ import satify.model.Result.*
 import satify.model.Status.COMPLETED
 import satify.model.cnf.CNF
 import satify.model.dpll.DecisionTree.{Branch, Leaf}
-import satify.model.dpll.PartialAssignment.{extractParAssignmentFromCnf, extractParAssignments}
 import satify.model.dpll.{Constraint, Decision, DecisionTree, OptionalVariable}
 import satify.model.{Assignment, Solution}
 import satify.update.solver.dpll.DpllDecision.decide
 import satify.update.solver.dpll.cnf.CNFSat.{isSat, isUnsat}
 import satify.update.solver.dpll.cnf.CNFSimplification.simplifyCnf
+import satify.update.solver.dpll.utils.PartialAssignmentUtils.*
+
 
 import scala.annotation.tailrec
 import scala.util.Random
